+++ src/mocks/datas/studyGroupDetail.ts
